# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'TiendaBackend API Documentation'
  description: 'Backend API del sistema de ventas Tienda. Módulo de autenticación y gestión de sesiones.'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtén tu token haciendo <b>POST /api/login</b> con tus credenciales. Luego pégalo aquí para probar los endpoints protegidos.'
security:
  -
    default: []
paths:
...
  /api/products/query:
    post:
      summary: 'Query products (POST)'
      operationId: queryProductsPOST
      description: "Supports two payload formats. Formato B (standard):\n`p.page` = row offset (0-indexed), `p.per_page`, `o.column`, `o.direction`, `w` = object.\nFormato A (compact):\n`p.r` = row offset, `p.s` = per_page, `o.field`, `o.type`, `w` = array of `{f, ao, v, lo}`."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  ok: true
                  code: 200
                  status: OK
                  message: 'Products retrieved.'
                  data:
                    items: []
                    totalCount: 0
                    summary:
                      - 0
                properties:
                  ok:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: 'Products retrieved.'
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        example: []
                      totalCount:
                        type: integer
                        example: 0
                      summary:
                        type: array
                        example:
                          - 0
                        items:
                          type: integer
      tags:
        - Inventory
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                p:
                  type: object
                  description: 'Pagination. `page`/`per_page` or `r`/`s`.'
                  example: []
                  properties: {  }
                f:
                  type: array
                  description: 'Fields to return.'
                  example:
                    - id
                    - name
                  items:
                    type: string
                o:
                  type: object
                  description: 'Order. `column`+`direction` or `field`+`type`.'
                  example: []
                  properties: {  }
                w:
                  type: object|array
                  description: 'Where filters. Object for simple equality, array for advanced operators.'
                  example: architecto
                totalCount:
                  type: boolean
                  description: 'Include total count (default true).'
                  example: true
...
  /api/inventory/movements:
    get:
      summary: 'List inventory movements (Kardex)'
      operationId: listInventoryMovementsKardex
      description: 'Supports filters: `p[page]`, `p[per_page]`, `f[]`, `o[column]`, `o[direction]`, `w[column]`, `totalCount`.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  ok: true
                  code: 200
                  status: OK
                  message: 'Inventory movements retrieved.'
                  data:
                    items: []
                    totalCount: 0
                    summary:
                      - 0
                properties:
                  ok:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: 'Inventory movements retrieved.'
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        example: []
                      totalCount:
                        type: integer
                        example: 0
                      summary:
                        type: array
                        example:
                          - 0
                        items:
                          type: integer
      tags:
        - Inventory
    post:
      summary: 'Register inventory movements'
      operationId: registerInventoryMovements
      description: "Registra uno o varios movimientos sobre variantes de un mismo producto y\nactualiza su existencia de forma atomica: si cualquier elemento de\n`movements` falla, no se aplica ninguno. Tipos que incrementan: `entry`,\n`return`, `cancel`. Tipos que disminuyen: `sale`, `adjustment`."
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  ok: true
                  code: 201
                  status: Created
                  message: 'Ajustes de inventario registrados correctamente.'
                  data:
                    movements:
                      -
                        id: 501
                        idProductVariant: 101
                        movementType: adjustment
                        quantity: 2
                        previousStock: 8
                        newStock: 6
                        referenceType: manual_adjustment
                        referenceId: null
                        notes: 'Ajuste por conteo físico'
                        idUser: 1
                        createdAt: '2026-08-12 21:45:00'
                    totalStock: 13
                properties:
                  ok:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 201
                  status:
                    type: string
                    example: Created
                  message:
                    type: string
                    example: 'Ajustes de inventario registrados correctamente.'
                  data:
                    type: object
                    properties:
                      movements:
                        type: array
                        example:
                          -
                            id: 501
                            idProductVariant: 101
                            movementType: adjustment
                            quantity: 2
                            previousStock: 8
                            newStock: 6
                            referenceType: manual_adjustment
                            referenceId: null
                            notes: 'Ajuste por conteo físico'
                            idUser: 1
                            createdAt: '2026-08-12 21:45:00'
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 501
                            idProductVariant:
                              type: integer
                              example: 101
                            movementType:
                              type: string
                              example: adjustment
                            quantity:
                              type: integer
                              example: 2
                            previousStock:
                              type: integer
                              example: 8
                            newStock:
                              type: integer
                              example: 6
                            referenceType:
                              type: string
                              example: manual_adjustment
                            referenceId:
                              type: string
                              example: null
                              nullable: true
                            notes:
                              type: string
                              example: 'Ajuste por conteo físico'
                            idUser:
                              type: integer
                              example: 1
                            createdAt:
                              type: string
                              example: '2026-08-12 21:45:00'
                      totalStock:
                        type: integer
                        example: 13
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  ok: false
                  code: 422
                  status: 'Unprocessable Entity'
                  message: 'La variante Negro / M no tiene existencia suficiente.'
                  data:
                    idProductVariant: 101
                    currentStock: 1
                    requestedQuantity: 2
                properties:
                  ok:
                    type: boolean
                    example: false
                  code:
                    type: integer
                    example: 422
                  status:
                    type: string
                    example: 'Unprocessable Entity'
                  message:
                    type: string
                    example: 'La variante Negro / M no tiene existencia suficiente.'
                  data:
                    type: object
                    properties:
                      idProductVariant:
                        type: integer
                        example: 101
                      currentStock:
                        type: integer
                        example: 1
                      requestedQuantity:
                        type: integer
                        example: 2
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idProduct:
                  type: integer
                  description: 'Producto al que pertenecen todas las variantes.'
                  example: 25
                idUser:
                  type: integer
                  description: 'Usuario que genera los movimientos (debe ser el de la sesión).'
                  example: 1
                referenceType:
                  type: string
                  description: 'Origen de los movimientos.'
                  example: manual_adjustment
                  nullable: true
                referenceId:
                  type: integer
                  description: 'Documento origen, si aplica.'
                  example: null
                  nullable: true
                notes:
                  type: string
                  description: 'Comentario general, aplicado a cada movimiento.'
                  example: 'Ajuste por conteo físico'
                  nullable: true
                movements:
                  type: array
                  description: 'Uno o más movimientos de inventario.'
                  example:
                    - architecto
                  items:
                    type: string
              required:
                - idProduct
                - idUser
                - movements
...